Refined namespace, scene and folder structure. Renamed BoundingRig to BoundingBoxRig.
