+++ Assets/MixedRealityToolkit/UX/Scripts/AppBar.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: e14fd115038007a4b8a7d7f9d61e2f9e
+folderAsset: yes
+timeCreated: 1519866245
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
