+++ Assets/MixedRealityToolkit/UX/Scripts/BoundingBoxes/BoundingBoxRig.cs.meta
